// src/app/interfaces/direccion.ts
export class Direccion {
  direccionId?: number;
  ubigeo?: string;
  personaId?: number;
  tipoVivienda?: string;
  tipoVia?: string;
  nombreVia?: string;
  numero?: string;
  manzana?: string;
  lote?: string;
  tipoSeccion?: string;
  numeroSeccion?: string;
  tipoZona?: string;
  nombreZona?: string;
  tipoSector?: string;
  nombreSector?: string;
  tipoDireccion?: string;
  codEstado?: string;
  correspondencia?: string;
}