// src/app/interfaces/acuerdo-pago.ts
export class AcuerdoPago {
  id?: number;
  asignacionId?: number;
  ejecutivoId?: number;
  socioId?: number;
  creditoId?: number;
  grupo?: number;
  tipoAcuerdo?: number;
  montoAcordado?: number;
  posibilidadPago?: string;
  fechaInicio?: string;
  horaIncio?: string;
  condicion?: string;
  descripcion?: string;
  cuota?: number;
  plazo?: number;
  fechaRegistro?: string;
  intervalo?: number;
}